move shared functions that do not depend on app components to shared directory
