+++ app/places-youll-go/src/features/counter/Counter.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react"
 
-import { useAppDispatch, useAppSelector } from "@/app/hooks"
+import { useAppDispatch, useAppSelector } from "@/shared/hooks"
 import styles from "./Counter.module.css"
 import {
   decrement,
