+++ app/places-youll-go/src/features/counter/counterSlice.ts
@@ -1,5 +1,5 @@
 import type { PayloadAction } from "@reduxjs/toolkit"
-import { createAppSlice } from "@/app/createAppSlice"
+import { createAppSlice } from "@/shared/createAppSlice"
 import type { AppThunk } from "@/app/store"
 import { fetchCount } from "./counterAPI"
