+++ app/places-youll-go/src/shared/hooks.ts
@@ -5,7 +5,7 @@
 // for importing and re-exporting the typed versions of hooks.
 /* eslint-disable @typescript-eslint/no-restricted-imports */
 import { useDispatch, useSelector } from "react-redux"
-import type { AppDispatch, RootState } from "./store"
+import type { AppDispatch, RootState } from "../app/store"
 
 // Use throughout your app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>()
